# bikesharing-rest-api/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Bikesharing REST API
  version: 0.0.1
tags:
  - name: test
    description: lorem ipsum
paths:
  /hello:
    get:
      tags:
        - test
      summary: Says hello on behalf of given user
      operationId: sayHello
      parameters:
        - name: from
          in: query
          description: blah
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            text/html:
              schema:
                type: string
                example: Hello World from Pavel!
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
  /stands:
    get:
      tags:
        - stands
      summary: Retrieves all stands currently in the system.
      operationId: retrieveStands
      responses:
        '200':
          description: Stands are retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StandDTO'
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /bikes/{bikeId}/location:
    parameters:
      - name: bikeId
        in: path
        required: true
        description: Unique bike identifier.
        schema:
          type: string
          format: uuid
    post:
      tags:
        - bikes
      summary: Moves bike with given ID to a new location.
      operationId: moveBike
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationDTO"
        required: true
      responses:
        '204':
          description: Bike location is updated.
        '400':
          description: Coordinates are not formatted correctly.
        '404':
          description: Bike does not exist.
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '409':
          description: Bike does not belong to any started ride.
        '500':
          description: Server error.
components:
  schemas:
    StandDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        location:
          $ref: '#/components/schemas/LocationDTO'
      required:
        - id
        - name
        - location
    LocationDTO:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
      required:
        - longitude
        - latitude